Déplace le slider avec sauvegarde auto dans un composant dédié
