+++ app.territoiresentransitions.react/src/referentiels/actions/avancement-detaille/ScoreAutoModal.tsx
