+++ app.territoiresentransitions.react/src/referentiels/actions/avancement-detaille/ScoreDetailleModal.tsx
