+++ app.territoiresentransitions.react/src/referentiels/actions/avancement-detaille/avancement-detaille.modal.tsx
