+++ app.territoiresentransitions.react/src/referentiels/actions/avancement-detaille/avancement-detaille.slider.tsx
